change main.py to only run when called directly not when imported by other parts; simple keras example of the same architecture, simple numpy example of the same general architecture
