+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__
 .vscode
 readme_adden.txt
+intro_tensorflow_example.py
